pyproject.toml
[tool.poetry]
name = "mdfy"
version = "0.1.0"
description = "A FastAPI-based web service that converts various document formats to Markdown"
authors = ["Jaigouk Kim <ping@jaigouk.kim>"]
license = "AGPL-3.0"
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.11.0,<3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
redis = "^5.0.6"
fastapi-cache2 = "^0.2.1"
python-magic = "^0.4.24"
magika = "^0.5.0"
pymupdf4llm = "^0.0.5"
returns = "^0.23.0"

[tool.poetry.dev-dependencies]
pytest = "^8.2.2"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
